Changed the the version of NVidia Docker that start_instance installs
Before this patch, start_instance installed the latest version of Docker but an old, incompatible version of NVidia Docker. This patch changes the version of NVidia Docker to the 1.0.1, which is compatible as of May 12th, 2017. However, the script will need to be updated whenever there is a new versionof NVidia Docker.

I tested this change by using the script to create a new instance and confirming that skynet-data and skynet-train run properly on that instance.